requirements.txt
APScheduler==3.5.3
attrs==19.1.0
autopep8==1.4.3
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.1.2
django-bulk-update==2.2.0
django-cors-headers==2.4.0
django-environ==0.4.5
djangorestframework==3.8.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
jsonschema==3.0.1
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
nbformat==4.4.0
numpy==1.15.4
pandas==0.23.4
parso==0.4.0
pickleshare==0.7.5
Pillow==5.3.0
plotly==3.9.0
prompt-toolkit==2.0.9
psutil==5.6.2
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
pycodestyle==2.5.0
Pygments==2.4.0
pyparsing==2.3.0
PyPDF2==1.26.0
pyrsistent==0.15.1
python-dateutil==2.7.5
pytz==2018.5
reportlab==3.5.9
requests==2.20.0
retrying==1.3.3
six==1.11.0
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.24
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.3
xlrd==1.2.0
